added animated sprite class
easily extensible in terms of adding new animations. simply add the triggers, define the new sequence in enum Animation_State, and ensure that the enum and the spritesheet are both ordered in the same sequence. currently uses a hacky magic number for frame count; an Animation_Sequence struct should be developed for storing that data, to enable variation in frame count and speed.

all entities still share a single texture. that should also be updated; perhaps a Texture_Manager class. (singleton?)