implemented #2, #8, #9, and #21
entities and animations are now handled more generically. the latter are also smoother, with direction-change transitions. enemy management now relies on a vector of prototype enemies, which should make spawning faster.

entities are organized along a hierarchy of Faction -> Type. animations are further subdivided by Direction. and it's all data-driven!