updated damage handling (close #49)
there is now a discrete struct, appropriately named Damage. it holds the projectile's type as well as the damage it deals.

Entity has been modified to use these in the (newly-renamed) takeDamage function.

in addition, a map of damage resistance values has been added to Entity. it is unusued beyond a debug statement in the Enemy constructor, and is waiting on database handling (see #50)