prevent repeated sound effect cascading
there is now a clock object associated with each sound type, in the new Sound_Context class. new sounds will not be added if the threshold hasn't been passed.

tangential issue, but: it's a little weird with sliders - the noise never stops when click dragging them. projected solution is to place the noise in the scroll class, and then calculate a mouse.x diff in ::update() (just remember to ignore it if it's OOB)